app/python/requirements.txt
flask
pyqrcode
pyzbar
Pillow